+++ ZendeskSell/Deals/DealActions.cs
@@ -28,7 +28,7 @@ namespace ZendeskSell.Deals {
             return (await _client.ExecuteAsync<ZendeskSellCollectionResponse<DealResponse>>(request, Method.GET)).Data;
         }
 
-        public async Task<ZendeskSellObjectResponse<DealResponse>> GetOneAsync(int id) {
+        public async Task<ZendeskSellObjectResponse<DealResponse>> GetOneAsync(long id) {
             var request = new RestRequest($"deals/{id}", Method.GET);
             return (await _client.ExecuteAsync<ZendeskSellObjectResponse<DealResponse>>(request, Method.GET)).Data;
         }
@@ -45,7 +45,7 @@ namespace ZendeskSell.Deals {
             return (await _client.ExecuteAsync<ZendeskSellObjectResponse<DealResponse>>(request, Method.POST)).Data;
         }
 
-        public async Task<ZendeskSellObjectResponse<DealResponse>> UpdateAsync(int id, DealRequest deal) {
+        public async Task<ZendeskSellObjectResponse<DealResponse>> UpdateAsync(long id, DealRequest deal) {
             Require.Argument("ContactID", deal.ContactID);
             Require.Argument("StageID", deal.StageID);
             Require.Argument("OwnerID", deal.OwnerID);
@@ -57,7 +57,7 @@ namespace ZendeskSell.Deals {
             return (await _client.ExecuteAsync<ZendeskSellObjectResponse<DealResponse>>(request, Method.PUT)).Data;
         }
 
-        public async Task<ZendeskSellDeleteResponse> DeleteAsync(int id) {
+        public async Task<ZendeskSellDeleteResponse> DeleteAsync(long id) {
             var request = new RestRequest($"deals/{id}", Method.DELETE);
             return (await _client.ExecuteAsync<ZendeskSellDeleteResponse>(request, Method.DELETE)).Data;
         }
